% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highcharts.R
\name{highcharts}
\alias{highcharts}
\title{A wrapper for Highcharts.js graphing library (http://api.highcharts.com/highcharts)}
\usage{
highcharts(data = NULL, x = NULL, y = NULL, group = NULL,
  type = "scatter", title = "Test", xAxis = list(title = list(text =
  "Fruit Eaten 1")), yAxis = list(title = list(text = "Fruit Eaten 2")),
  width = NULL, height = NULL, chartOpts = list(), creditsOpts = list(),
  exportingOpts = list(), plotOptions = list())
}
\arguments{
\item{data}{A data.frame or data.table object, which must be in a long
format.}

\item{x}{Name of the x-axis variable.}

\item{y}{Name of the y-axis variable.}

\item{group}{Name of the group variable. Default is NULL.}

\item{type}{Type of the graph to be produced. See
http://api.highcharts.com/highcharts for the possibilities.}

\item{title}{Title of the graph}

\item{xAxis}{A named list of options to be passed to the Highcharts API (see
http://api.highcharts.com/highcharts for details)}

\item{yAxis}{A named list of options to be passed to the Highcharts API (see
http://api.highcharts.com/highcharts for details)}

\item{width}{Width of the graph...}

\item{height}{Height of the graph...}

\item{chartOpts}{A named list of options to be passed to the Highcharts API (see
http://api.highcharts.com/highcharts for details)}

\item{creditsOpts}{A named list of options to be passed to the Highcharts API (see
http://api.highcharts.com/highcharts for details)}

\item{exportingOpts}{A named list of options to be passed to the Highcharts API (see
http://api.highcharts.com/highcharts for details)}

\item{plotOptions}{A named list of options to be passed to the Highcharts API (see
http://api.highcharts.com/highcharts for details)}

\item{dataList}{A list of data.tables, one for each series to be plotted.}
}
\description{
This function is a wrapper around the Highcharts.js library. This function is
a wrapper around the Highstock.js library. Currently, this is still an early
development version, so there are still quite many glitches.
}

