% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highstocks.R
\name{highstocks}
\alias{highstocks}
\title{A wrapper for Highstock.js graphing library (http://api.highcharts.com/highstock)}
\usage{
highstocks(data = NULL, x = NULL, y = NULL, group = NULL,
  type = "scatter", title = "Test", xAxis = list(title = list(text =
  "Fruit Eaten 1")), yAxis = list(title = list(text = "Fruit Eaten 2")),
  width = NULL, height = NULL, chartOpts = list(), creditsOpts = list(),
  exportingOpts = list(), plotOptions = list())
}
\arguments{
\item{data}{A data.frame or data.table object, which must be in a long
format.}

\item{x}{Name of the x-axis variable. `highstocks` function expects this to be
formatted a s a date.}

\item{y}{Name of the y-axis variable.}

\item{group}{Name of the group variable. Default is NULL.}

\item{type}{Type of graph to be produced. Currently, this wrapper supports
only a line option.}

\item{title}{Title of the graph.}

\item{xAxis}{A named list of options to be passed to the Highstock API (see
http://api.highcharts.com/highstock for details)}

\item{yAxis}{A named list of options to be passed to the Highstock API (see
http://api.highcharts.com/highstock for details)}

\item{width}{Width of the graph...}

\item{height}{Height of the graph...}

\item{chartOpts}{A named list of options to be passed to the Highstock API (see
http://api.highcharts.com/highstock for details)}

\item{creditsOpts}{A named list of options to be passed to the Highstock API (see
http://api.highcharts.com/highstock for details)}

\item{exportingOpts}{A named list of options to be passed to the Highstock API (see
http://api.highcharts.com/highstock for details)}

\item{plotOptions}{A named list of options to be passed to the Highstock API (see
http://api.highcharts.com/highstock for details)}
}
\description{
This function is a wrapper around the Highstock.js library. Currently, this
is still an early development version, so there are still quite many glitches.
}

