% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highcharts.R
\name{renderHighcharts}
\alias{renderHighcharts}
\title{Widget render function for use in Shiny}
\usage{
renderHighcharts(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

