% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highstocks.R
\name{renderHighstocks}
\alias{renderHighstocks}
\title{Widget render function for use in Shiny}
\usage{
renderHighstocks(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

